requirements.txt
hydra-core
numpy~=1.21.4
pandas~=1.3.4
pyspark~=3.2.0
wget~=3.2
matplotlib~=3.4.3
google-cloud-bigquery
pandas-gbq
scipy~=1.7.2
scikit-learn~=1.0.1
omegaconf~=2.1.1
black